More changes suggested by splint.
git-svn-id: svn://svn.code.sf.net/p/muhttpd/code/current@39 7e1f86b9-3cf0-0310-9d02-db70efa590c4

+++ config.c
@@ -359,8 +359,8 @@ static struct muhttpd_config *parse_config_line(char *line,
 		return config;
 	} else if(!strcmp(tok, "index")) {
 		while((tok = get_next_token(&line))) {
-			p = realloc(config->index,
-				(config->indices + 1) * sizeof(char*));
+			p = realloc(config->index, (size_t)
+				((config->indices + 1) * sizeof(char*)));
 			if(!p) {
 				fputs("ERROR: Error adding index", stderr);
 				config = NULL;
